refactor frontend js
